First version

+++ notePlayer.hpp
@@ -1,7 +1,7 @@
 #include "Arduino.h"
 #include "constants.hpp"
 #pragma once
-void PlayNote(int);
+void PlayNote(float);
 void BadTune();
 void GoodTune();
 void picoTone(float,unsigned long int);
